add brochures page in contentController Fixes #18
