+++ src/Controller/UsersController.php
@@ -2,91 +2,57 @@
 
 namespace App\Controller;
 
+// use generals
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
-use App\Entity\User;
-use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\String\Slugger\SluggerInterface;
-use App\Form\UserType;
-use App\Repository\UserRepository;
-use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Doctrine\ORM\EntityManagerInterface;
+
+// use for User
+use App\Entity\User;
+use App\Form\UserType;
+use App\Repository\UserRepository;
+
+
 
 
 
 class UsersController extends AbstractController
 {
+    /**
+     * class for manage the users
+     *
+     * @return Response
+     * @method  index() to display the list of users
+     * @method  new() to add a new user
+     */
+
     #[Route('/users', name: 'app_users')]
     public function index(): Response 
     {
+        /**
+         * Function to display the list of users
+         * 
+         * @Route("/users", name="app_users")
+         * @return Response
+         * 
+         */
+
         $user = $this->getUser();
         return $this->render('users/index.html.twig', [
             'controller_name' => 'UsersController',
             'user' => $user,
         ]);
+    }
 
 
-        
-    }
-    #[Route('/users/list', name: 'app_users_list')]
-    public function list(EntityManagerInterface $em)
-    {
-        $users = $em->getRepository(User::class)->findAll();
-    
-        return $this->render('users/list.html.twig', [
-            'users' => $users,
-        ]);
-    }
  
-
-    #[Route('/user/files', name: 'app_user_new')]
-    public function new(Request $request, SluggerInterface $slugger): Response
-    {
-        $user = new User();
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            /** @var UploadedFile $brochureFile */
-            $brochureFile = $form->get('brochure')->getData();
-
-            // this condition is needed because the 'brochure' field is not required
-            // so the PDF file must be processed only when a file is uploaded
-            if ($brochureFile) {
-                $originalFilename = pathinfo($brochureFile->getClientOriginalName(), PATHINFO_FILENAME);
-                // this is needed to safely include the file name as part of the URL
-                $safeFilename = $slugger->slug($originalFilename);
-                $newFilename = $safeFilename.'-'.uniqid().'.'.$brochureFile->guessExtension();
-
-                // Move the file to the directory where brochures are stored
-                try {
-                    $brochureFile->move(
-                        $this->getParameter('brochures_directory'),
-                        $newFilename
-                    );
-                } catch (FileException $e) {
-                    // ... handle exception if something happens during file upload
-                }
-
-                // updates the 'brochureFilename' property to store the PDF file name
-                // instead of its contents
-                $user->setBrochureFilename($newFilename);
-                
-            }
-
-            // ... persist the $user variable or any other work
-
-            return $this->redirectToRoute('app_users');
-        }
-
-        return $this->render('users/files/addFiles.html.twig', [
-            'form' => $form,
-        ]);
-    }
+ 
+    
+    
 
     // Display the files of a user
     #[Route('/user/files/{slug}', name: 'app_user_files')]
@@ -103,6 +69,6 @@ class UsersController extends AbstractController
     $safeFilename = $originalName;
     $file->move($targetDirectory, $safeFilename);
     
-}
-    
+    }
+
 }
