+++ src/Service/PictureService.php
@@ -12,11 +12,24 @@ class PictureService
      * @return string
      */
     private $params;
-    public function getPicture(ParameterBagInterface $params)
+    public function __construct(ParameterBagInterface $params)
     {
-        return 'picture';
+        $this->params = $params;
+
+    }
+    public function add($UploadedFile $picture, ?string $folder = "", ?int $width = null, ?int $height = null, ?string $name = null)
+    {
+        $fileName = md5(uniqid()).'.'.$file->guessExtension();
+        $file->move($this->params->get('upload_directory'), $fileName);
+    }
+    {
+        $fileName = md5(uniqid()).'.'.$file->guessExtension();
+        $file->move($this->params->get('upload_directory'), $fileName);
+    }
+    {
+        $file->move($this->params->get('upload_directory'), $fileName);
     }
 }
-```
+
 
 ?>
